requirements.txt
streamlit
faiss-cpu
numpy
sentence-transformers
langchain
langchain-community
pypdf
groq
